Add weak minimax functionality
After the "go" command, tell the Search() function search to the depth that it is called with. I am not sure that the minimax is working correctly, though, because I told it to cout "Making white move" and "Making black move" when it is examining each move. These outputs appear to be correct if it searches to depth 2, but not 4 or 6. The last line of output is the nodecount; the number of positions it searched. This number appears to be correct at 2 and 4 and maybe 6 plies. We'll see if I can get it to work by the next commit.