Fix the eval problem
This was the problem with the alpha-beta search! Apparently, the compiler used external linkage on the White-Score and Black_Score vars in the header file. I had enough of the global variables, so I put them in the Evaluate_Position function and lo and behold, the search method worked! There are a few more problems with the search, and I will continue to work on it. This single commit has done the most good for this project!