Make a Move smaller
Change the stack of temp moves from Bitboards to ints. This has the effect of a pretty good speedup because of the less memory to be passed around in the search function. However, there have been a couple of crashes and a couple of stalls but they seem to be fixed; I will keep an eye on them.