Make iterative deepening into root search
Now the engine uses a root search. This mainly has the effect of allowing more effective leveraging of the pv and the "currmove" commands. In short, the root search is mainly to beautify the output to the GUI. I have a copy of the engine before this commit, so I can revert to that if This does not work out