Fix castling being legal in wrong places in the tree
We didn't keep track of castling rights in the tree, so the engine would make a castling move and then still think castling was possible later on down the tree. The only time the castling rights were correct was at ply 0. This would cause some hallucinations since the engine would incorrectly think castling was possible and result in incorrect scores deep in the tree.